Name and currency fix (#21)
* Remove name validation

* Revert "Remove name validation"

This reverts commit 9725af322b498992acbba2f00d7124bdec0242e8.

* Fix email enter screen layout for long name support

* Fix validation conditions for name

* Fix email enter screen layout

* Add price precision and update price formatter

* Update version to 1.0.1 and build number to 5

* Add target specific localization string files

* Fix approved receipt description text

* Add display of target specific text for approved receipt description

* Fix onboarding text

+++ WhiteLabel-iOS/Common/Data/Models/ProductModel.swift
@@ -15,8 +15,8 @@ class ProductModel: BaseDataModel {
   var name = ""
   var description = ""
   var specs = ""
-  var price = 0
-  var discountedPrice = 0
+  var price = 0.0
+  var discountedPrice = 0.0
 
   // MARK: - Overridden methods
   override func update(with jsonData: JSON) {
@@ -27,7 +27,7 @@ class ProductModel: BaseDataModel {
     name = jsonData["name"].stringValue
     description = jsonData["description"].stringValue
     specs = jsonData["specs"].stringValue
-    price = jsonData["price"].intValue
-    discountedPrice = jsonData["discounted_price"].intValue
+    price = jsonData["price"].doubleValue
+    discountedPrice = jsonData["discounted_price"].doubleValue
   }
 }
